gh-105481: move Python/opcode_metadata.h to Include/internal/pycore_opcode_metadata.h
