+++ Include/internal/pycore_opcode_metadata.h
